pyproject.toml
[project]
name = "expense-tracking-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.122.0",
    "mysql-connector-python>=9.5.0",
    "pandas>=2.3.3",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
]
